requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.4
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
cloudscraper==1.2.71
chardet==5.2.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
greenlet==3.0.3
sentence-transformers==5.2.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
markdown==3.5.1
